+++ packages/viz/test/helpers/pdfDoc.js
@@ -15,6 +15,7 @@
  * == BSD2 LICENSE ==
  */
 
+import sinon from 'sinon';
 import MemoryStream from 'memorystream';
 
 export default class Doc {